fix: fix lint error, delete obsolete mocks from /api/deck
