--- app/api/deck/all/route.ts
@@ -1,13 +1,0 @@
-import { NextResponse } from "next/server"
-import { readDecks } from "../decks-json"
-
-export async function GET(request: Request) {
-  const decks = readDecks()
-
-  return NextResponse.json({ data: decks })
-}
-
-// export type SearchResponse = {
-//   data: Scry.Card[] | null
-//   error?: string
-// }
