--- app/api/deck/decks.json
@@ -1,143 +1,0 @@
-[
-  {
-    "id": "1",
-    "name": "Glittering arts & auras",
-    "champions": [
-      {
-        "id": "d9713032-2956-4564-b5f5-2dd16245a4e6",
-        "image_uri": "https://cards.scryfall.io/art_crop/front/d/9/d9713032-2956-4564-b5f5-2dd16245a4e6.jpg?1572489608"
-      },
-      {
-        "id": "1100b898-31a8-4fdf-a54f-a1470ec032f3",
-        "image_uri": "https://cards.scryfall.io/art_crop/front/1/1/1100b898-31a8-4fdf-a54f-a1470ec032f3.jpg?1592485685"
-      }
-    ],
-    "cards": [
-      ["d9713032-2956-4564-b5f5-2dd16245a4e6", 4],
-      ["1100b898-31a8-4fdf-a54f-a1470ec032f3", 4],
-      ["9e7fb3c0-5159-4d1f-8490-ce4c9a60f567", 2],
-      ["769cddf0-2078-456b-a622-057571464fe7", 2],
-      ["a10c1407-d397-4caa-b7b7-e7d91ffd4ee9", 3],
-      ["f6d5314b-4994-4cd0-a680-12e56be7c1e2", 4],
-      ["b42e1359-c79d-4848-a0f1-3f274e0bf0d9", 2],
-      ["8cd5cc66-2ade-4142-9269-7d9905b029e5", 1],
-      ["d5f69e24-6b5c-4a9a-af5b-1ce88efdc0aa", 1],
-      ["e0cb828c-a19b-4e6d-853f-85e5ea7cadf8", 1],
-      ["611497d2-83df-4c69-bbeb-6a807c686bc4", 1],
-      ["30c22eb9-5056-4c0b-a5d8-41e09161eb40", 1],
-      ["53a7cccd-b445-46e3-b70c-afc3bb2de85d", 1],
-      ["b9da673d-7cc0-4435-b5a5-5098630f7712", 1],
-      ["2e658cb2-11c5-4b59-a91c-f26ac28e82af", 1],
-      ["b0247ee3-7799-43c2-9617-35b3edc91f22", 1]
-    ]
-  },
-  {
-    "id": "2",
-    "name": "Knights",
-    "champions": [
-      {
-        "id": "330cc452-4382-401d-9432-ac27ae6e27ad",
-        "image_uri": "https://cards.scryfall.io/art_crop/front/3/3/330cc452-4382-401d-9432-ac27ae6e27ad.jpg?1572490205"
-      },
-      {
-        "id": "c7d5e394-8e41-442e-ae97-a478a61e1b9d",
-        "image_uri": "https://cards.scryfall.io/art_crop/front/c/7/c7d5e394-8e41-442e-ae97-a478a61e1b9d.jpg?1572489629"
-      }
-    ],
-    "cards": [
-      ["c7d5e394-8e41-442e-ae97-a478a61e1b9d", 2],
-      ["99083707-2152-42c0-b5c3-b4f97ec20190", 1],
-      ["7791f46e-f772-4d3f-824e-52b4fc721b58", 1],
-      ["4b2bde2d-e5df-407e-993c-85880dbb6045", 2],
-      ["0001f1ef-b957-4a55-b47f-14839cdbab6f", 2],
-      ["3d1a3fec-39de-4223-9da2-22749a58cd62", 2],
-      ["4cddb2d2-d813-4b83-a592-380ba4edf54f", 1],
-      ["050a0817-2e9c-4d98-974c-2d3e5c37e1a2", 1],
-      ["52dd633f-9a07-47df-adb1-36013fc2f43a", 1],
-      ["2c3c992a-70b5-4d3c-9a96-93c3365691ac", 2],
-      ["fe0a63bb-dd94-429f-aa9b-21f3d1c53ae5", 1],
-      ["c5f6c745-e46a-42eb-8eca-b7b74ab1245e", 1],
-      ["da6f21a8-3dd0-42af-8a93-84f98968c781", 1],
-      ["2d6139d3-5397-4403-9c1e-312c11a7542b", 1],
-      ["330cc452-4382-401d-9432-ac27ae6e27ad", 2],
-      ["9173ffda-1d3b-4dab-8dcb-de44717de464", 1],
-      ["e73d8a84-2c0d-423c-89c7-71de0af9e1ac", 1],
-      ["f82541f2-b17c-45b4-87ff-f9b46d23578c", 2],
-      ["a808868f-aea8-4651-9357-85a4d7b4f290", 1],
-      ["17755d1b-3a56-4362-a534-85b35ceb1802", 1],
-      ["6bcfe06b-3bb5-434e-acf4-4bfa8a94141d", 1],
-      ["6cc73d16-5ed7-4104-91f6-0997a2080e2e", 1],
-      ["c52d66db-5570-48a1-99cf-e0417517747b", 1],
-      ["0e96f3dc-dc58-4034-8d3d-3ae74fa64562", 1],
-      ["a3d13d84-01e4-4429-93db-e5afff811527", 1],
-      ["33a78207-fd76-4112-a257-54a25da6f818", 1],
-      ["080a249a-df47-4769-bb63-0d8ab3f2467c", 1],
-      ["1b2b198d-493c-4d6c-bfb6-e842728522f6", 1],
-      ["9a8b1fea-3c5d-43d2-b4d2-e8938f3f7b9c", 1],
-      ["0c3f372d-259d-4a31-9491-2d369b3f3f8b", 1],
-      ["06e8d715-025f-4578-b427-e401318a9c58", 1],
-      ["27425f2e-e0b2-489d-877d-8257d2026bfd", 1],
-      ["39c6e952-97b5-435c-a7fd-6271ddc23200", 1],
-      ["aa9d0279-4fe0-4228-a82f-bc3cd91500df", 1],
-      ["6716f46c-806d-4b3e-8a1c-fd6dcedacf8e", 1],
-      ["8a665794-513f-4f78-92c9-1844ec27c79c", 1],
-      ["853dbf04-7165-4728-b96c-c9e4ff4d3491", 1],
-      ["f6601056-af08-4239-97d5-5e11597fce18", 1],
-      ["9c8a3119-ce9b-4fe5-aea3-871a7eee216e", 1],
-      ["166ee7fc-a0f7-49af-85bf-d5b36bb2bff4", 1],
-      ["4348513e-4a03-4e06-98e0-ac6c0dfb013d", 1],
-      ["b537c4e3-88d1-430d-b4c8-16e0716d1927", 1],
-      ["a33add37-379d-4a90-9c04-529dff676986", 1]
-    ]
-  },
-  {
-    "id": "3",
-    "name": "Unearthed",
-    "champions": [
-      {
-        "id": "a59f4e5c-fdc7-485f-aadb-2a71b3701dcc",
-        "image_uri": "https://cards.scryfall.io/art_crop/front/a/5/a59f4e5c-fdc7-485f-aadb-2a71b3701dcc.jpg?1562201638"
-      }
-    ],
-    "cards": [
-      ["4a835b7e-cd63-486d-8b58-2f443f686e32", 2],
-      ["a59f4e5c-fdc7-485f-aadb-2a71b3701dcc", 4],
-      ["c6db664e-f0c4-4e60-add7-89b9f210cbc7", 1],
-      ["b547513d-8b69-41cd-84c9-4b08b6426f1d", 2],
-      ["b1f73271-b70f-40ae-be64-f8de7805923a", 1],
-      ["a960acea-3404-4834-9296-d6715add49af", 1],
-      ["db87a2be-4490-45af-89d7-540549d1d6af", 1]
-    ]
-  },
-  {
-    "id": "4",
-    "name": "msk deck #1",
-    "cards": [
-      ["9a539a23-8383-4525-82dd-acfe1d219fe9", 1],
-      ["2b9bb2f3-a7c7-4e8f-ba39-44e6acd9240b", 1],
-      ["7873c1f9-572c-4740-82f8-cf3cbc7318d0", 1],
-      ["c5f6c745-e46a-42eb-8eca-b7b74ab1245e", 1],
-      ["330cc452-4382-401d-9432-ac27ae6e27ad", 1],
-      ["9173ffda-1d3b-4dab-8dcb-de44717de464", 1],
-      ["9514ce0d-8d61-4cce-b2ad-8f417869d04e", 2],
-      ["a59f4e5c-fdc7-485f-aadb-2a71b3701dcc", 2],
-      ["888b4b76-6af9-4479-8cfd-cde3b8694fe5", 1],
-      ["333406d5-abcc-4629-a33b-395d0662ba1b", 1],
-      ["c6db664e-f0c4-4e60-add7-89b9f210cbc7", 1],
-      ["b547513d-8b69-41cd-84c9-4b08b6426f1d", 2],
-      ["49c37d70-799d-464a-972c-11a78ae31272", 1],
-      ["b62abd0c-ec3e-45d7-989d-da269812aeef", 2],
-      ["fb662d58-ae50-423d-b1c5-abd45baca12b", 2],
-      ["7f9a5267-eb90-43bc-bf92-fcfb06821bae", 1],
-      ["4dc774b4-3f70-4351-b1b8-8a0193cb3a50", 1],
-      ["77972745-8689-4733-9a9d-39ae9d6273a2", 1],
-      ["7c33d20d-8d1c-4cf7-a7ad-3dab9758fafe", 1]
-    ],
-    "champions": [
-      {
-        "id": "77972745-8689-4733-9a9d-39ae9d6273a2",
-        "image_uri": "https://cards.scryfall.io/art_crop/front/7/7/77972745-8689-4733-9a9d-39ae9d6273a2.jpg?1592516666"
-      }
-    ]
-  }
-]
